Optimize publish code path with dummy timer
Replaces `maybe_delay_first` with `delay_first`.  Modfies the behavior
so it ensures that there is always one timer active, even when there
are no delayed messages in the exchange.  This timer is not used for
anything except book-keeping, so we set it to expire far in the future
(1 hour) so it doesn't gobble up resources for no reason.

Previously the `erlang:read_timer(CurrTimer) == false` code path could
be hit in two scenarios:
- No timer had been set.  The message currently being added is the
  only one in the delay queue.
- The timer has already expired.  The delay queue is full of messages
  ready for publication and the exchange has not yet caught up.
In the former case, the exchange should start a new timer.  In the
latter, the exchange is far behind and starting a new timer will only
waste time.  The new invariant makes the first of those scenarios
impossible.

Updates the publication code path to take advantage of this change.
Since the first of those two scenarios is no longer possible, the
publication code path is no longer responsible for starting timers.
(It can still *replace* a timer if the current timer is set too far
in the future, but it can't create a new one from nothing.)