Improve tracking of expired and unset timers
Adds explicit tracking for the case where no timer is set.  In that
case, the timer is assigned to the `not_set` atom.  This allows us to
distinguish between the cases where no timer is set and the case where
the timer was set but has expired.

Previously, the `internal_delay_message` function was unable to tell the
difference between these two cases.  This is a problem because the
action it should take in either case is different.  If the timer was
never set, the enqueue code path must start one.  If the timer has
expired and will be processed soon, then starting a new timer would be a
waste of time.  The previous solution was to always start the timer,
since that does not affect correctness.

Now that we can distinguish the two scenarios, the enqueue path can
handle each case independently.  This makes enqueing significantly
faster in the case where enqueuing coincides with messages becoming
ready for delivery.