Don't get stuck on index key with no messages
When the delayed_message gen_server encounters the next timestamp in the
index table and there are no messages in the main table associated, it
would loop forever, as this timestamp will stay the "first" key in the
index table.

This commit makes sure the expired index entry gets deleted even if
there are no associated messages.

This situation can happen if the node is abruptly stopped in the middle
of delaying a messsage (after writing to the index table but before
writing to the main table).